# HDDS-9328. Speed up testListStatusIteratorOnPageSize